+++ src/main/scala/stellar/sdk/model/PaymentSigningRequest.scala
@@ -1,7 +1,8 @@
 package stellar.sdk.model
 
 import okhttp3.HttpUrl
-import stellar.sdk.util.ByteArrays
+import okio.ByteString
+import stellar.sdk.util.{ByteArrays, DoNothingNetwork}
 import stellar.sdk.{KeyPair, PublicKey}
 
 import scala.Option
@@ -14,13 +15,18 @@ import scala.Option
  * @param memo a memo to attach to the transaction
  * @param callback the uri to post the transaction to after signing
  * @param message an optional message for displaying to the user
+ * @param networkPassphrase the passphrase of the target network, if it's not the public/main network
+ * @param signature a domain and signature that proves the validity of this signing request
+ * @see See [[https://github.com/stellar/stellar-protocol/blob/master/ecosystem/sep-0007.md#operation-pay|SEP-0007]] for full specification
  */
 case class PaymentSigningRequest(
   destination: PublicKey,
   amount: Option[Amount] = None,
   memo: Memo = NoMemo,
   callback: Option[HttpUrl] = None,
-  message: Option[String] = None
+  message: Option[String] = None,
+  networkPassphrase: Option[String] = None,
+  signature: Option[DomainSignature] = None
 ) {
 
   message.foreach(m => require(m.length <= 300, "Message must not exceed 300 characters"))
@@ -43,7 +49,10 @@ case class PaymentSigningRequest(
       "callback" -> callback.map(c => s"url:$c"),
       "memo" -> memoEncoded.map(_._1),
       "memo_type" -> memoEncoded.map(_._2),
-      "msg" -> message
+      "msg" -> message,
+      "network_passphrase" -> networkPassphrase,
+      "origin_domain" -> signature.map(_.originDomain),
+      "signature" -> signature.map(_.signature.base64())
     ).filter(_._2.isDefined).foldLeft(Map.empty[String, String]) {
       case (m, (k, Some(v))) => m.updated(k, v)
       case (m, _) =>            m
@@ -97,8 +106,15 @@ object PaymentSigningRequest {
 
     val message = Option(httpUrl.queryParameter("msg"))
 
+    val networkPassphrase = Option(httpUrl.queryParameter("network_passphrase"))
+
+    val signature = for {
+      domain <- Option(httpUrl.queryParameter("origin_domain"))
+      sigBytes <- Option(httpUrl.queryParameter("signature")).map(ByteString.decodeBase64)
+    } yield DomainSignature(domain, sigBytes)
+
     Option(httpUrl.queryParameter("destination")).map(KeyPair.fromAccountId).map { destination =>
-      PaymentSigningRequest(destination, amount, memo, callback, message)
+      PaymentSigningRequest(destination, amount, memo, callback, message, networkPassphrase, signature)
     }
     .getOrElse(throw new IllegalArgumentException(s"Invalid url: [url=$url]"))
   }