+++ src/main/scala/stellar/sdk/model/TransactionSigningRequest.scala
@@ -18,7 +18,8 @@ import scala.concurrent.{ExecutionContext, Future}
  * @param pubkey the public key associated with the signer who should sign
  * @param message an optional message for displaying to the user
  * @param networkPassphrase the passphrase of the target network, if it's not the public/main network
- * @see See [[https://github.com/stellar/stellar-protocol/blob/master/ecosystem/sep-0007.md#operation-tx SEP-0007 for full specification]]
+ * @param signature a domain and signature that proves the validity of this signing request
+ * @see See [[https://github.com/stellar/stellar-protocol/blob/master/ecosystem/sep-0007.md#operation-tx|SEP-0007]] for full specification
  */
 case class TransactionSigningRequest(
   transaction: SignedTransaction,
@@ -86,7 +87,9 @@ case class TransactionSigningRequest(
       "callback" -> callback.map(c => s"url:$c"),
       "pubkey" -> pubkey.map(_.accountId),
       "msg" -> message,
-      "network_passphrase" -> networkPassphrase
+      "network_passphrase" -> networkPassphrase,
+      "origin_domain" -> signature.map(_.originDomain),
+      "signature" -> signature.map(_.signature.base64())
     ).filter(_._2.isDefined).foldLeft(Map.empty[String, String]) {
       case (m, (k, Some(v))) => m.updated(k, v)
       case (m, _) =>            m
@@ -158,9 +161,14 @@ object TransactionSigningRequest {
         // TODO (jem) - The transaction objects should take a strongly typed network passphrase object.
         .map(p => new DoNothingNetwork(p))
 
+    val signature = for {
+      domain <- Option(httpUrl.queryParameter("origin_domain"))
+      sigBytes <- Option(httpUrl.queryParameter("signature")).map(ByteString.decodeBase64)
+    } yield DomainSignature(domain, sigBytes)
+
     Option(httpUrl.queryParameter("xdr"))
         .map(SignedTransaction.decodeXDR(_)(network.getOrElse(PublicNetwork)))
-        .map(TransactionSigningRequest(_, form, callback, pubKey, message, network.map(_.passphrase)))
+        .map(TransactionSigningRequest(_, form, callback, pubKey, message, network.map(_.passphrase), signature))
       .getOrElse(throw new IllegalArgumentException(s"Invalid url: [url=$url]"))
   }